Cleanup deep hash code
